+++ tcms/rpc/api/product.py
@@ -3,8 +3,8 @@
 from django.forms.models import model_to_dict
 from modernrpc.core import rpc_method
 
+from tcms.management.forms import ProductForm
 from tcms.management.models import Product
-from tcms.rpc.api.forms.management import ProductForm
 from tcms.rpc.decorators import permissions_required
